fix: resolve all javac compilation warnings (63 warnings) and update todo.md
- BlobStorage: close/delete/getOrOpen -> async + .join()
- getSoundType(level, pos, null) вместо deprecated getSoundType()
- @SuppressWarnings для неизбежных deprecation/removal/unchecked
- DefaultSessionLayer: добавлен break в fallthrough
- Intra16x16PredictionBuilder: явный каст (byte) для lossy conversion
- ResourceBlockDeviceData: close() throws Exception -> IOException
- InternetCardSpec: defineListAllowEmpty -> defineList + @SuppressWarnings
- NetworkSwitchBlockEntity: rawtypes/unchecked, getList() вместо каста
- SwitchPortManager: List<Map> -> List<Map<String, ?>>
- PMD ruleset: удалены несуществующие exclude-правила
- build.gradle.kts: -Xlint:-classfile для BaselineIgnore
- todo.md: переписан под актуальное состояние

+++ src/main/java/li/cil/oc2/jcodec/codecs/h264/decode/Intra16x16PredictionBuilder.java
@@ -117,7 +117,7 @@ public final class Intra16x16PredictionBuilder {
 
         for (int i = 0; i < pred.length; i++)
             for (int j = 0; j < pred[i].length; j++)
-                pred[i][j] += s0;
+                pred[i][j] += (byte) s0;
     }
 
     public static int lumaDCPredSAD(final boolean leftAvailable, final boolean topAvailable, final byte[] leftRow, final byte[] topLine, final int x,
